+++ src/api/antrol-auto/antrolAuto.controller.ts
@@ -67,6 +67,37 @@ router.get("/update-task/:limit/task/:task_id", async (req: Request, res: Respon
     }
 })
 
+
+router.get("/update-task-backdate/:limit/task/:task_id", async (req: Request, res: Response, next: NextFunction) => {
+    try {
+        const limit: number     = parseInt(req.params.limit, 10);
+        const task_id: number   = parseInt(req.params.task_id, 10);
+        const dataUpdateTaskNow: any = await updateTask(limit, task_id, true);
+
+        if (dataUpdateTaskNow.length > 0) {
+            res.status(200).json({
+                metadata: {
+                    code: 200,
+                    msg: "Operation completed successfully!",
+                },
+                response: {
+                    dataUpdateTaskNow,
+                },
+            });
+            console.log(dataUpdateTaskNow);
+        } else {
+            res.status(200).json({
+                metadata: {
+                    code: 200,
+                    msg: "Data tidak tersedia!",
+                },
+            });
+        }
+    } catch (error: any) {
+        next(error.message.replace(/\n/g, " "));
+    }
+})
+
 router.get("/update-task-fisio/:limit/task/:task_id", async (req: Request, res: Response, next: NextFunction) => {
     try {
         const limit: number     = parseInt(req.params.limit, 10);